+++ components/post-navigator.js
@@ -1,14 +1,41 @@
-import { usePreviousPost } from './use-previous-post';
-import { useNextPost } from './use-next-post';
+import { usePostNavigor } from './use-post-navigator';
+import Link from 'next/link';
 
 export const PostNavigator = ({ page, posts }) => {
-  const previousPost = usePreviousPost(page, posts);
-  const nextPost = useNextPost(page, posts);
+  const {
+    length,
+    indexNumber,
+    nextPageTitle,
+    previousPageTitle,
+    nextPageId,
+    previousPageId,
+    nextPagePath,
+    previousPagePath,
+  } = usePostNavigor(page, posts);
 
   return (
     <div className="post-navigator">
-      <div>{previousPost}</div>
-      <div>{nextPost}</div>
+      <div>
+        {indexNumber === 0 ? null : (
+          <div>
+            Previous: <br />
+            <Link key={previousPageId} href={`/posts${previousPagePath}`}>
+              {`👈 ${previousPageTitle}`}
+            </Link>
+          </div>
+        )}
+      </div>
+      <div>
+        {indexNumber + 1 === length ? null : (
+          <div>
+            Next:
+            <br />
+            <Link key={nextPageId} href={`/posts${nextPagePath}`}>
+              {`${nextPageTitle} 👉`}
+            </Link>
+          </div>
+        )}
+      </div>
 
       <style jsx>
         {`
